+++ gherkin-messages/src/main/java/io/cucumber/core/gherkin/messages/GherkinMessagesExamples.java
@@ -14,13 +14,15 @@ final class GherkinMessagesExamples implements Node.Examples {
     private final io.cucumber.messages.types.Examples examples;
     private final List<Example> children;
     private final Location location;
+    private final Node parent;
 
-    GherkinMessagesExamples(io.cucumber.messages.types.Examples examples) {
+    GherkinMessagesExamples(Node parent, io.cucumber.messages.types.Examples examples) {
+        this.parent = parent;
         this.examples = examples;
         this.location = GherkinMessagesLocation.from(examples.getLocation());
         AtomicInteger row = new AtomicInteger(1);
         this.children = examples.getTableBody().stream()
-                .map(tableRow -> new GherkinMessagesExample(tableRow, row.getAndIncrement()))
+                .map(tableRow -> new GherkinMessagesExample(this, tableRow, row.getAndIncrement()))
                 .collect(Collectors.toList());
     }
 
@@ -43,6 +45,11 @@ final class GherkinMessagesExamples implements Node.Examples {
     public Optional<String> getName() {
         String name = examples.getName();
         return name.isEmpty() ? Optional.empty() : Optional.of(name);
+    }
+
+    @Override
+    public Optional<Node> getParent() {
+        return Optional.of(parent);
     }
 
 }